Refactoring: class TTrack.NameDialog separated as TrackNameDialog
