+++ src/org/opensourcephysics/cabrillo/tracker/TTrack.java
@@ -58,7 +58,7 @@ public abstract class TTrack implements Interactive,
   protected static JCheckBox skippedStepWarningCheckbox;
   protected static JButton closeButton;
   protected static boolean skippedStepWarningOn = true;
-  protected static NameDialog nameDialog;
+  protected static TrackNameDialog nameDialog;
   protected static int nextID = 1;
   protected static HashMap<Integer, TTrack> activeTracks = new HashMap<Integer, TTrack>();
   protected static FontRenderContext frc
@@ -2901,63 +2901,6 @@ public abstract class TTrack implements Interactive,
 			}
 		}
 	}
-  
-  /**
-   * A dialog used to set the name of a track.
-   */
-  protected static class NameDialog extends JDialog {
-  	
-  	JLabel nameLabel;
-  	JTextField nameField;
-  	TTrack target;
-  	TrackerPanel trackerPanel;
-  	
-  	// constructor
-  	NameDialog(TrackerPanel panel) {  		
-  		super(panel.getTFrame(), null, true);
-  		trackerPanel = panel;
-      setDefaultCloseOperation(WindowConstants.DO_NOTHING_ON_CLOSE);
-      addWindowListener(new WindowAdapter() {
-        public void windowClosing(WindowEvent e) {
-        	String newName = nameField.getText();
-        	if (target != null) 
-        		trackerPanel.setTrackName(target, newName, true);
-        }
-      });
-      nameField = new JTextField(20);
-      nameField.addActionListener(new ActionListener() {
-        public void actionPerformed(ActionEvent e) {
-        	String newName = nameField.getText();
-        	if (target != null) 
-        		trackerPanel.setTrackName(target, newName, true);
-        }
-      });
-      nameLabel = new JLabel();
-      JToolBar bar = new JToolBar();
-      bar.setFloatable(false);
-      bar.add(nameLabel);
-      bar.add(nameField);
-      JPanel contentPane = new JPanel(new BorderLayout());
-      contentPane.add(bar, BorderLayout.CENTER);
-      setContentPane(contentPane);
-  	}
-  	
-    /**
-     * Sets the track.
-     * 
-     * @param track the track
-     */
-  	void setTrack(TTrack track) {
-  		target = track;
-      // initial text is current track name
-  		FontSizer.setFonts(this, FontSizer.getLevel());
-    	setTitle(TrackerRes.getString("TTrack.Dialog.Name.Title")); //$NON-NLS-1$
-    	nameLabel.setText(TrackerRes.getString("TTrack.Dialog.Name.Label")); //$NON-NLS-1$
-    	nameField.setText(track.getName());
-    	nameField.selectAll();
-    	pack();
-  	}
-  }
 
   /**
    * Returns an ObjectLoader to save and load data for this class.
@@ -3123,9 +3066,9 @@ public abstract class TTrack implements Interactive,
     }
   }
   
-  protected static NameDialog getNameDialog(TTrack track) {
+  protected static TrackNameDialog getNameDialog(TTrack track) {
   	if (nameDialog==null && track.trackerPanel!=null) {
-      nameDialog = new NameDialog(track.trackerPanel);
+      nameDialog = new TrackNameDialog(track.trackerPanel);
       Dimension dim = Toolkit.getDefaultToolkit().getScreenSize();
       int x = (dim.width - nameDialog.getBounds().width) / 2;
       int y = (dim.height - nameDialog.getBounds().height) / 2;
